+++ handtrack/demo/package.json
@@ -9,10 +9,12 @@
     "node": ">=8.9.0"
   },
   "dependencies": {
-    "@tensorflow/tfjs": "^1.3.2"
+    "@tensorflow-models/handtrack": "file:../",
+    "@tensorflow/tfjs-core": "^1.3.2",
+    "@tensorflow/tfjs-converter": "^1.3.2"
   },
   "scripts": {
-    "watch": "cross-env NODE_ENV=development parcel index.html --no-hmr --open ",
+    "watch": "cross-env NODE_ENV=development parcel index.html --no-hmr ",
     "build": "cross-env NODE_ENV=production parcel build index.html --public-url ./",
     "lint": "eslint .",
     "link-local": "yalc link @tensorflow-models/handtrack"
